requirements.txt
asgiref==3.9.2
certifi==2025.8.3
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-elasticsearch-dsl==8.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
elastic-transport==8.17.1
elasticsearch==8.19.1
elasticsearch-dsl==8.15.4
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
